Kill off the legacy dialog
If the SystemUI is down for the count, chances are that there are bigger problems
to worry about. If we're paying attention, reviewing things like we're supposed to
the SystemUI should never be down.

Also this causes a hiccup in that if SystemUI is restarted (which we added the option)
the legacy power menu shows up every time. That's janky, but that's not the reason for
removing this. I just don't think we should have two different dialogs. We're committing
to the new one!

Change-Id: I9642bfe58668edc706bdd3e5e70522eb4b712160
Signed-off-by: Genkzsz11 <Genkzsz11@gmail.com>